Refactor to use the new variable schema
In this patch, we move to using the config information under
neerali_ceph_config instead of <cluster-name> as it may not
be available across. Likewise, neerali_ceph_config is kept as
a golden copy and modifications are done only to variables in
scope of the task

Adding host to the cluster tasks are now included in the main
flow.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>