Improve reimage failure and conf generation
ipmitool is required for reimage operation hence added the
package to be installed. Likewise ipmi access information is
required for all the systems.

Modifications are made to the way we are generating the conf
file for teuthology. It is simplified to accept more key/value
pairs without modifying the code.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>