Cleanup playbook
This playbook contains the required tasks for cleaning up the
Ceph cluster. The inventory passed to it must have information
about all the _admin nodes and the controller node.

The _admin nodes are used to gather the fsid and disable orch.
While the controller which has passwordless access to all the
nodes is used to zap the nodes.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>