Add support for dynamic repos
In CI runs, there is a scenario where the list of repos are static
as well as dynamic. The static repos pertain to the RHEL based
repos whereas the dynamic repo information pertain to a Ceph
build.

Hence, this commit adds support of dynamic list of repos via the
variable `neerali_repo_setup_repos_extras`

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>