Switch to generic wait till service ready
This commit modifies the existing osd_wait_until tasks to a
generic one that can be leveraged by other services. It
introduces service_type variable to support this generalization

It also allows for customization of sleep (delay) and number of
retries. Using these variables the user can increase or decrease
the time by which the service must be ready.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>