Fix unit test execution in case of wrong labels
Read-only option works when volume bind happens on a Mac system
however, RH is a bit more stricter leading to permission denied.
This prevents us from executing the test.

Option 1 is to ensure proper labels are applied.
Option 2 is to switch from ro to Z

'Z' instructs podman to change the label and mark it as private
unshared label only for the current container.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>