Fix reimage RemoteDisconnected error
With the recent changes to urllib, we are consistently hitting
an issue with attempting a reimage. When teuthology is installed
using pip i.e. git+https protocol then the installed urllib3
package is the latest.

Hence, we are moving to installing via the repo instead. This
commit modifies these steps.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>