Fix incorrect variable name
Undefined variable error was thrown as the dict did not have a
key with the name service_name. The previous task prepares the
map having the key as name but the value is of service_name.

This commit fixes the undefined variable error by aligning the
key names being used.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>