Fix listeners workflow to iterate
When there are multiple subsystems for the same group, the
listeners task was not being executed for the second set of
subsystems due to the presence of the exected file.

The last part of the subsystem is different hence suffix it
with the log file should help avoid this issue.

Signed-off-by: Pragadeeswaran Sathyanarayanan <pragadeeswaran.sathyanarayanan@ibm.com>